Increase default box size to 25GB
Setting the disk space to 10GB by default means that you can't import
the default set of databases for all of the apps without using fdisk
to increase your partition.

This commit provides a sane default for all of the base data that we
import for all of the apps -- at the moment.