Upgrade veewee and remove vagrant
veewee no longer needs Vagrant to run. Which is good, because most of us are
now using binary installs of Vagrant and `bundle exec vagrant --version` no
longer returns the correct one.

Per the veewee repo/site we can just call `bundle exec veewee vbox`. It has
been upgraded from the version that was previously pinned in order to
support `veewee export`.

Everything else appears to work the same.

It would be nice to replace all this with Packer and also build VMware
Fusion boxes someday. But not today.