Build boxes for Ubuntu 14.04.1 (Trusty)
Because why wouldn't you?

No seriously, just for once if we can't be ahead of the curve of OS
releases, it would be nice to stay at least close to the curve.

To get this building, all I actually had to do was make the
puppetlabs-release package install dependent on the OS release - which
was pretty easy. Everything else worked as-is.

I'm aware that this isn't the prettiest Makefile in the world, full of
unnecessary duplication, but I left it like that as I barely understand
Make and I bet I'm not alone in this. At least this way most people
could follow what it is doing.

I've tested the boxes output by confirming there were no errors in the
build and I've tested them in our puppet repo to the point where puppet
runs, but will of course not run successfully as we have missing
packages for this release. But hey, it's closer than we were yesterday!