Fix Packer errors
The converted VeeWee configuration (see previous commit) failed Packer's
validator. Fix this using `packer fix`, which then shows another error
which we'll tackle next.

    matt➜~/govuk/boxes(packer)» packer validate output/template.json
    Failed to create build 'virtualbox':

    No builder returned for name: virtualbox
    matt➜~/govuk/boxes(packer)» packer fix output/template.json > output/temp.json
    matt➜~/govuk/boxes(packer✗)» mv output/temp.json output/template.json
    matt➜~/govuk/boxes(packer✗)» packer validate output/template.json
    Failed to parse template: 2 error(s) occurred:

    * provisioner 1: build 'virtualbox' not found for override
    * provisioner 1: build 'vmware' not found for override
    matt➜~/govuk/boxes(packer✗)»

The main changes here appear to be the sort order of the hash keys and
the name of the VirtualBox builder, whose name changed from `virtualbox`
to `virtualbox-iso` in Packer version 0.5.0 ([changelog]).

Note that the version of Packer I'm using is `0.7.1`.

[changelog]: https://github.com/mitchellh/packer/blob/master/CHANGELOG.md#050-12302013