Fix Packer validation errors
In the previous commit, `packer fix` amended the names of the Packer
builders in use, e.g. `virtualbox` became `virtualbox-iso`. It seems
`packer fix` missed the builder names in the `override` hash, so we
change them here to fix the error.

This configuration now validates successfully:

    matt➜govuk/boxes/output(packer✗)» packer validate template.json
    Template validated successfully.

I was able to successfully build a Virtualbox image using this
configuration by running:

    packer build -only=virtualbox-iso template.json

...although I cannot test that using Vagrant until I add a Vagrant
post-processor for Packer.