Use Packer variable to reduce duplication
Packer does not provide an easy way to reduce duplication between
templates; it does not support YAML for example (YAML anchors could
serve this purpose) and I'd rather not use a YAML-to-JSON preprocessor.

It does allow for user-defined variables that can be used in the JSON
template however, so removes duplication where possible using variables.

Note that we don't replace `boot_command` with a variable because it
contains variables itself which wouldn't be evaluated if we were to make
it into a variable.