Delete VeeWee scripts and update README
This commit removes the now deprecated VeeWee configuration and updates
the `README` explaining how to use Packer.

Note that this commit removes the `Makefile`; the relevant task left in
the Makefile was the one to upload the box templates to Amazon S3, but
I've moved this command to `README.md`. I prefer that to leaving it in
the Makefile as I think it's good to expose the underlying command in
this case.

We also move the contents of `output/` into the repository's root
directory and rename `template.json` to `templates.json` to make clear
that it contains multiple templates.